Add "Convert to material preset" feature
A new feature that allows you to convert a model's material to the preset one. Useful for converting P5R materials or adding a blue glow effect for custom personas.
To use it, right click on the material parent folder, select Convert to -> Material preset, and choose the preset.
Currently the batch conversion isn't working because I'm dumb lole