Add new menu option to copy an animation set of split GAPs
Adds an option that allows the user to select a split a/b/c combat animations GAP and select a directory with multiple models, this will create a per model copy of the GAP files and fix the bone IDs for every model.

The GAPs are automatically placed in a battle subfolder where the models are located.

There is a prompt asking the user if the GAP ID is 51 or 52 (certain characters like Makoto have GAP ID 52 too).