Enhanced the parsing capability of GFDLibrary for GNF Texture format.
1.
The GetSurfaceFormat function now supports parsing BC7 format when the DDS format is DX10

2.
This issue has been resolved in the current enhancement, which ensures that other attributes are not overwritten when writing attribute values due to the shared mWord (32-bit value) in the header field of GNF Texture format.

3.
Fixed the issue of insufficient space in the source or destination array in the DoSwizzle function of Array.Copy (only occurring in some GNF files).

4.
Enhanced the capability of decoding GNF texture to DDS format in the DecodeToDDS function.